requirements.txt
langchain
langchain-core
langchain-openai
langchain-tavily
langchainhub
black
python-dotenv
requests
